Remove unused parts in Megatron-LM code and add copyright notice
