+++ cacheflow/model_executor/parallel_utils/__init__.py
@@ -1,6 +1,5 @@
 import cacheflow.model_executor.parallel_utils.parallel_state
 import cacheflow.model_executor.parallel_utils.tensor_parallel
-import cacheflow.model_executor.parallel_utils.utils
 
 # Alias parallel_state as mpu, its legacy name
 mpu = parallel_state
@@ -8,5 +7,4 @@ mpu = parallel_state
 __all__ = [
     "parallel_state",
     "tensor_parallel",
-    "utils",
 ]
