+++ cacheflow/model_executor/parallel_utils/tensor_parallel/__init__.py
@@ -17,15 +17,12 @@ from .mappings import (
 )
 
 from .random import (
-    checkpoint,
     get_cuda_rng_tracker,
     model_parallel_cuda_manual_seed,
 )
 
 from .utils import (
     split_tensor_along_last_dim,
-    split_tensor_into_1d_equal_chunks,
-    gather_split_1d_tensor,
 )
 
 __all__ = [
@@ -45,11 +42,8 @@ __all__ = [
     "scatter_to_tensor_model_parallel_region",
     "scatter_to_sequence_parallel_region",
     # random.py
-    "checkpoint",
     "get_cuda_rng_tracker",
     "model_parallel_cuda_manual_seed",
     # utils.py
     "split_tensor_along_last_dim",
-    "split_tensor_into_1d_equal_chunks",
-    "gather_split_1d_tensor",
 ]
