+++ cacheflow/model_executor/parallel_utils/tensor_parallel/mappings.py
@@ -1,3 +1,5 @@
+# Copyright 2023 The CacheFlow team.
+# Adapted from https://github.com/NVIDIA/Megatron-LM/blob/main/megatron/core/tensor_parallel/mappings.py
 # Copyright (c) 2022, NVIDIA CORPORATION. All rights reserved.
 
 import torch
